adk_multiagent/creative_director/requirements.txt
google-adk==1.18.0
a2a-sdk==0.2.16
uvicorn
fastapi
httpx==0.28.1
google-genai
python-dotenv 
aiohttp==3.12.15 
requests 
mermaid-python==0.1 
nest-asyncio==1.6.0
google-cloud-bigquery
google-cloud-aiplatform
google-cloud-storage
google-auth
toolbox-core==0.5.0
google-cloud-logging